refactor(#337): SelectBox 컴포넌트 위치 변경
