+++ src/app/class/[id]/_components/Apply.tsx
@@ -4,12 +4,14 @@ import { useRouter } from 'next/navigation';
 import { useState, useRef } from 'react';
 import { toast } from 'react-toastify';
 import { useClickAway } from 'react-use';
-import SelectBox from './apply/SelectBox';
+import SelectBox from './SelectBox';
 import { ApplyButton } from '@/components/Button';
 import { IClassSchedule, IDateTime } from '@/types/class';
 import { formatDateTime, applyScheduleFilter } from '@/utils/parseUtils';
 
-const ReservationItem = dynamic(() => import('./apply/ReservationItem'));
+const ReservationItem = dynamic(
+  () => import('@/components/Apply/ReservationItem'),
+);
 
 interface ApplyProps {
   id: string | number;
